requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2
django-admin-thumbnails==0.2.5
django-ckeditor==6.0.0
django-currencies==0.10.1
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.5.0
django-js-asset==1.2.2
django-mptt==0.12.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==2.10
jsonfield==2.1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
phonenumbers==8.13.9
Pillow==8.2.0
psycopg2==2.9.1
python-dateutil==2.8.2
python-gettext==4.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.4
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
